fix: change credit amount type from Decimal to int

+++ deployment/migrations/versions/0035_a1b2c3d4e5f6_add_credit_balances_table.py
@@ -23,7 +23,7 @@ def upgrade() -> None:
         'credit_balances',
         sa.Column('id', sa.BigInteger(), autoincrement=True),
         sa.Column('address', sa.String(), nullable=False),
-        sa.Column('amount', sa.DECIMAL(), nullable=False),
+        sa.Column('amount', sa.BigInteger(), nullable=False),
         sa.Column('ratio', sa.DECIMAL(), nullable=True),
         sa.Column('tx_hash', sa.String(), nullable=True),
         sa.Column('token', sa.String(), nullable=True),
